==> explore_figure/validate/4_total_cell_proportion.Rmd <==
---
title: "03_Make_ylim_and_change_to_cell_proportion"
output: html_document
---

```{r setup, include=FALSE}
library(Seurat)
library(here)
library(tidyverse)
library(cowplot)
library(knitr)
library(Matrix)
library(future)
library(future.apply)
```

refquery <- readRDS(here("explore_figure/covid_validation_3/RDS/05_Integrate_cell_type_main.RDS"))


```{r}
unique(refquery$Diagnosis)
```

# subset by Diagnosis 
```{r}
refquery_IPF <- subset(x = refquery, subset = Diagnosis == "IPF")
refquery_COVID <- subset(x = refquery, subset = Diagnosis == "COVID")
refquery_Control <- subset(x = refquery, subset = Diagnosis == "Control")
refquery_Control_B <- subset(x = refquery, subset = Diagnosis == "Control B")
refquery_covid_severe <- subset(x = refquery, subset = Diagnosis == "COVID severe")
refquery_covid_moderate <- subset(x = refquery, subset = Diagnosis == "COVID moderate")


```

# 1. IPF
```{r}
cells_IPF <- refquery_IPF@meta.data
cells_IPF %>% select(cell_type_main)

cell_count_df1 <- cells_IPF %>% 
  select(cell_type_main) %>% 
  group_by(cell_type_main) %>% 
  summarise(count_cell = n(), .groups = "drop")

sum_cell_df1 <- cell_count_df1 %>% 
  group_by(cell_type_main) %>% 
  summarise(sum_cell = sum(count_cell))

sum1 <- cell_count_df1 %>% left_join(sum_cell_df1, by = 'cell_type_main') 
as.data.frame(sum1)
head(sum1)
```

```{r}
ggplot(sum1,
       aes(x = cell_type_main, 
           y = count_cell,
           color = cell_type_main,
           fill = cell_type_main)) + 
  geom_bar(stat = "identity", width = 0.7) + 
  labs(x = "IPF Cell-Type", y = "Cell-Percent") + 
  facet_grid(~cell_type_main,
             scales = "free",
             space = "free") + 
  theme_classic() + 
  labs(title = "IPF Cell proportion") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 8, face = 'bold'),
        plot.title = element_text(hjust = 0.5, size = 10, face = 'bold'), legend.position = 'bottom') + 
  coord_cartesian(ylim = c(0,17000)) + 
  NoLegend()

ggsave(here('explore/Figure_workflow/cell_proportion/IPF_cellpropportion.pdf')) 
```

# covid
```{r}
cells_IPF <- refquery_COVID@meta.data
cells_IPF %>% select(cell_type_main)

cell_count_df1 <- cells_IPF %>% 
  select(cell_type_main) %>% 
  group_by(cell_type_main) %>% 
  summarise(count_cell = n(), .groups = "drop")

sum_cell_df1 <- cell_count_df1 %>% 
  group_by(cell_type_main) %>% 
  summarise(sum_cell = sum(count_cell))

sum1 <- cell_count_df1 %>% left_join(sum_cell_df1, by = 'cell_type_main') 
as.data.frame(sum1)
head(sum1)
```

```{r}
ggplot(sum1,
       aes(x = cell_type_main, 
           y = count_cell,
           color = cell_type_main,
           fill = cell_type_main)) + 
  geom_bar(stat = "identity", width = 0.7) + 
  labs(x = "IPF Cell-Type", y = "Cell-Percent") + 
  facet_grid(~cell_type_main,
             scales = "free",
             space = "free") + 
  theme_classic() + 
  labs(title = "IPF Cell proportion") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 8, face = 'bold'),
        plot.title = element_text(hjust = 0.5, size = 10, face = 'bold'), legend.position = 'bottom') + 
  coord_cartesian(ylim = c(0,17000)) + 
  NoLegend()

ggsave(here('explore/Figure_workflow/cell_proportion/IPF_cellpropportion.pdf')) 
```

# covid severe
```{r}
cells_IPF <- refquery_covid_severe@meta.data
cells_IPF %>% select(cell_type_main)

cell_count_df1 <- cells_IPF %>% 
  select(cell_type_main) %>% 
  group_by(cell_type_main) %>% 
  summarise(count_cell = n(), .groups = "drop")

sum_cell_df1 <- cell_count_df1 %>% 
  group_by(cell_type_main) %>% 
  summarise(sum_cell = sum(count_cell))

sum1 <- cell_count_df1 %>% left_join(sum_cell_df1, by = 'cell_type_main') 
as.data.frame(sum1)
head(sum1)
```

```{r}
ggplot(sum1,
       aes(x = cell_type_main, 
           y = count_cell,
           color = cell_type_main,
           fill = cell_type_main)) + 
  geom_bar(stat = "identity", width = 0.7) + 
  labs(x = "IPF Cell-Type", y = "Cell-Percent") + 
  facet_grid(~cell_type_main,
             scales = "free",
             space = "free") + 
  theme_classic() + 
  labs(title = "IPF Cell proportion") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 8, face = 'bold'),
        plot.title = element_text(hjust = 0.5, size = 10, face = 'bold'), legend.position = 'bottom') + 
  coord_cartesian(ylim = c(0,17000)) + 
  NoLegend()

ggsave(here('explore/Figure_workflow/cell_proportion/IPF_cellpropportion.pdf')) 
```
covid moderate
# covid
```{r}
cells_IPF <- refquery_covid_moderate@meta.data
cells_IPF %>% select(cell_type_main)

cell_count_df1 <- cells_IPF %>% 
  select(cell_type_main) %>% 
  group_by(cell_type_main) %>% 
  summarise(count_cell = n(), .groups = "drop")

sum_cell_df1 <- cell_count_df1 %>% 
  group_by(cell_type_main) %>% 
  summarise(sum_cell = sum(count_cell))

sum1 <- cell_count_df1 %>% left_join(sum_cell_df1, by = 'cell_type_main') 
as.data.frame(sum1)
head(sum1)
```

```{r}
ggplot(sum1,
       aes(x = cell_type_main, 
           y = count_cell,
           color = cell_type_main,
           fill = cell_type_main)) + 
  geom_bar(stat = "identity", width = 0.7) + 
  labs(x = "IPF Cell-Type", y = "Cell-Percent") + 
  facet_grid(~cell_type_main,
             scales = "free",
             space = "free") + 
  theme_classic() + 
  labs(title = "IPF Cell proportion") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 8, face = 'bold'),
        plot.title = element_text(hjust = 0.5, size = 10, face = 'bold'), legend.position = 'bottom') + 
  coord_cartesian(ylim = c(0,17000)) + 
  NoLegend()

ggsave(here('explore/Figure_workflow/cell_proportion/IPF_cellpropportion.pdf')) 
```

# control b
```{r}
cells_IPF <- refquery_Control_B@meta.data
cells_IPF %>% select(cell_type_main)

cell_count_df1 <- cells_IPF %>% 
  select(cell_type_main) %>% 
  group_by(cell_type_main) %>% 
  summarise(count_cell = n(), .groups = "drop")

sum_cell_df1 <- cell_count_df1 %>% 
  group_by(cell_type_main) %>% 
  summarise(sum_cell = sum(count_cell))

sum1 <- cell_count_df1 %>% left_join(sum_cell_df1, by = 'cell_type_main') 
as.data.frame(sum1)
head(sum1)
```

```{r}
ggplot(sum1,
       aes(x = cell_type_main, 
           y = count_cell,
           color = cell_type_main,
           fill = cell_type_main)) + 
  geom_bar(stat = "identity", width = 0.7) + 
  labs(x = "IPF Cell-Type", y = "Cell-Percent") + 
  facet_grid(~cell_type_main,
             scales = "free",
             space = "free") + 
  theme_classic() + 
  labs(title = "IPF Cell proportion") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 8, face = 'bold'),
        plot.title = element_text(hjust = 0.5, size = 10, face = 'bold'), legend.position = 'bottom') + 
  coord_cartesian(ylim = c(0,17000)) + 
  NoLegend()

ggsave(here('explore/Figure_workflow/cell_proportion/IPF_cellpropportion.pdf')) 
```




```{r}
# COVID
cells_IPF <- refquery_COVID@ meta.data
cells_IPF %>% select(cell_type_submain)

cell_count_df1 <- cells_IPF %>% 
  select(cell_type_submain, cell_type_main) %>% 
  group_by(cell_type_submain, cell_type_main) %>% 
  summarise(count_cell = n(), .groups = "drop")

sum_cell_df1 <- cell_count_df1 %>% 
  group_by(cell_type_submain, cell_type_main) %>% 
  summarise(sum_cell = sum(count_cell))

sum1 <- cell_count_df1 %>% left_join(sum_cell_df1, by = 'cell_type_submain') 
as.data.frame(sum1)
head(sum1)

ggplot(sum1,
       aes(x = cell_type_submain, 
           y = count_cell,
           color = cell_type_main.x,
           fill = cell_type_main.x)) + 
  geom_bar(stat = "identity", width = 0.7) + 
  labs(x = "COVID Cell-Type", y = "Cell-Count") + 
  facet_grid(~cell_type_main.x,
             scales = "free",
             space = "free") + 
  theme_classic() + 
  labs(title = "COVID Cell proportion") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 8, face = 'bold'),
        plot.title = element_text(hjust = 0.5, size = 10, face = 'bold'), legend.position = 'bottom') + 
  coord_cartesian(ylim = c(0,17000)) + 

  NoLegend()

ggsave(here('explore/Figure_workflow/cell_proportion/COVID_cellpropportion.pdf')) 


# Control 
cells_IPF <- refquery_Control@ meta.data
cells_IPF %>% select(cell_type_submain)

cell_count_df1 <- cells_IPF %>% 
  select(cell_type_submain, cell_type_main) %>% 
  group_by(cell_type_submain, cell_type_main) %>% 
  summarise(count_cell = n(), .groups = "drop")

sum_cell_df1 <- cell_count_df1 %>% 
  group_by(cell_type_submain, cell_type_main) %>% 
  summarise(sum_cell = sum(count_cell))

sum1 <- cell_count_df1 %>% left_join(sum_cell_df1, by = 'cell_type_submain') 
as.data.frame(sum1)
head(sum1)

ggplot(sum1,
       aes(x = cell_type_submain, 
           y = count_cell,
           color = cell_type_main.x,
           fill = cell_type_main.x)) + 
  geom_bar(stat = "identity", width = 0.7) + 
  labs(x = "Control Cell-Type", y = "Cell-Count") + 
  facet_grid(~cell_type_main.x,
             scales = "free",
             space = "free") + 
  theme_classic() + 
  labs(title = "Control Cell proportion") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 8, face = 'bold'),
        plot.title = element_text(hjust = 0.5, size = 10, face = 'bold'), legend.position = 'bottom') + 
  coord_cartesian(ylim = c(0,17000)) + 

  NoLegend()

ggsave(here('explore/Figure_workflow/cell_proportion/Control_cellpropportion.pdf')) 
```
